EPMRPP-90325 || Code Review fix - 1 : Reorganize project structure by adding 'organization' folder in 'pages' and updating file paths
