+++ app/src/pages/organization/organizationProjectsPage/emptyProjectsState/emptyProjectsState.scss
