+++ app/src/pages/organization/organizationProjectsPage/header/projectsPageHeader.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-.container {
+.projects-page-header-container {
   padding: 16px 32px;
   border-bottom: 1px solid $COLOR--e-100;
   background: $COLOR--bg-000;
