+++ app/src/pages/organization/organizationProjectsPage/organizationProjectsPage.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-.organization-page-container{
+.organization-projects-container{
   display: flex;
   flex-direction: column;
   height: 100%;
